+++ database/migrations/2024_12_11_181130_add_sort_to_page_page_images.php
@@ -1,0 +1,25 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration {
+    /**
+     * Run the migrations.
+     */
+    public function up(): void {
+        Schema::table('page_page_image', function (Blueprint $table) {
+            $table->integer('sort')->default(0);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void {
+        Schema::table('page_page_image', function (Blueprint $table) {
+            $table->dropColumn('sort');
+        });
+    }
+};